feat: move shareable components to SelectComponents file
